Fixed bug in the analytical squeezed approx.
In SONG, all bispectra are computed as <X_l1 Y_l2 Z_l3>. Therefore the field
X has to be associated with l1, Y with l3 and Z with l3. If the association
is broken, the Fisher estimator is broken too, as it connect each bispectrum
product to its covariance based on this link. In writing the analytic
approximation I wrongly associated X with l3 and this gave unreasonably
high values for the Fisher matrix elements of the T+E bispectrum (F~2000).
Now that the problem is fixed, I correctly get F~3, which is much more in
line with the intrinsic bispectrum results.