First implementation of the CMB-lensing bispectrum.
I have implemented the CMB-lensing bispectrum in Eq. 4.4 and 4.6 of
Lewis, Challinor & Hanson (2011, http://uk.arxiv.org/abs/1101.2234).
I have tested it against the implementation of CAMB. While I find the
shape of the Fisher matrix estimator vs the smallest l to be the same,
my lensing bispectrum is about 2 times larger for TTT and 16 times
larger for T+E. If we account for a factor 2 coming from not having
implemented yet the lensing variance in the Fisher estimator (see
page 14 of the paper and onward), we are still off a factor 8 for T+E.

Part of the problem is related to the fact that the lensing potential
C_l's in CLASS are different than those in CAMB for l>100, with the
former being larger. Also, CAMB uses the lensed C_l's rather than
the normal one, both in the estimator covariance and in the formula
for the CMB-lensing bispectrum. I will now send an e-mail to Julien
concerning the lensing potential, and implement a version of SONG
with the lensed C_l's.