Implemented lensed C_l's in SONG.
It is a bit rudimentary. I am calling perturbations.c and transfer.c two times,
once for the usual computation and the second time exclusively to compute
the lensed C_l's with an l_max increased by ppr->delta_l_max. It is a bit
slow for simple bispectra such as the analytical ones and the separable
templates, but it works. An alternative approach would be to run transfer.c
only once but keep track of two different k_max's. This implies modifying
stuff in the second-order modules too, so I leave it for future work.

Now, I will refine the Fisher matrix integrator. The isw-lensing bispectrum
is a bitch because it oscillates quickly, so I am forced to take a lot of points
(>250 per direction). This density in l, when including also the intrinsic
bispectrum, would make the computation last forever. Therefore, I will be
implementing a similar approach to CAMB's for the simple bispectra: still store
them in pbi->bispectra, but produce them analytically in the Fisher matrix estimator.
For example, for the mesh_2d interpolation I would end up avoiding completely the
function mesh_int for the analytical bispectra, which probably takes as much time
as computing the analytical bispectra. The advantage is that now I am only interpolating
in the l1 dimension, just like CAMB, which is (should be?) smooth after summing over
l2 and l3.