First implementation of the lensing variance.
I have implemented Eq. 5.25 of Lewis, Challinor & Hanson 2011
(http://uk.arxiv.org/abs/1101.2234) to account for the noise
contribution given by lensing in the Fisher matrix estimator.
The current results are wrong, because the equation relies on
the Fisher matrix as a function of the smallest multipole, l3,
which is computed in SONG/CLASS only for ppr->l_linstep==1. I
am gonna fix this in the next commit.