First working implementation of lensing variance.
We match CAMB within 3%. Problem is that the lensing variance function
still relies on the Fisher matrix as a function of the smallest l, so that
the results can be trusted only for ppr->linstep==1. Will now try to amend
this by either allowing storing F(l3) ppr->linstep!=1 or by changing the
lensing variance function to use l1 instead of l3.