Fully implemented lensing variance in Fisher estimator.
I have included the lensing variance as in Sec. 5 of http://uk.arxiv.org/abs/1101.2234.
With respect to Eq. 5.35 of that reference, I factor out the C_l1^{i\phi}C_l1^{j\phi}
at the denominator in the square brackets. I am not sure whether this is legit (the
square bracket is matrix-inverted after all!), but it is done also in CAMB.
Will ask Antony for confirmation.

Now SONG's matches CAMB's at the 5% level in the Fisher matrix, and at the
2% level in the S/N. The match improves further if I include the lensing
variance in SONG only up to where the CMB-lensing bispectrum is computed, that is, up to
l=MAX(pbi->l_max_lensing_corrT, pbi->l_max_lensing_corrE), as it is done in CAMB.
CAMB probably does this because those are the scales where the linear C_l^{\phi\phi}
is not accurate anymore. On the other hand, I think C_l^{\phi\phi} is still a good
approximation to the true one, and is also numerically stable in CLASS. Therefore,
I will compute the lensing variance all the way to l_max.