Substantially improved 2D mesh interpolation.
Thanks to the proper 2D mesh interpolation (rather than the 3D one
we were using before), I am now interpolating the two largest l’s
(l2 and l3) rather than the two smallest ones (l1 and l2). By doing so,
I end up interpolating smooth 2D slices (which look almost as 1D slices
for small l1 because of the triangular condition) instead of
bad-behaved ones (with an almost-divergent behaviour in the squeezed
corners of the l1-l2 plane).

By doing so, SONG’S Fisher matrix as a function of l1, F_l1, does not show
strange wiggly features anymore. Furthermore, the interpolation convergence
improved substantially at l1<200, so that now SONG is also quicker.
Now all F_l1 curves produced by SONG are smoother and possess a shape more
similar to the analytical approximation.

There still is the problem of SONG’s Fisher matrix being larger than the analytical
approximation for l1<100 (see green curve vs red curve in the 3rd plot). I found out
that such discrepancy almost disappears when I exclude reionisation (see fifth plot,
which is a low precision run up to l_max=2000). I am now trying to understand why
reionisation would produce such a big difference, but I suspect it is related to the
fact that we do not include it in the second-order transfer functions.

As a final (positive) note, I have restricted the estimator only to l1>200 to isolate
the signal of the subhorizon effects, which is not affected by lensing variance.
This would be the area under the magenta bumps at l~900 in the third plot.
With l_max=3000, I find that these effects alone (which are not included in the
analytical approximation) provide a signal to noise of about S/N=1.5. It is important
to note, however, that this signal has yet to be lensed. As Antony pointed out, this
might be achieved using Eq. 2.11 in http://arxiv.org/abs/1201.1010, if we assume that
most of the signal comes from (mildly) squeezed configurations. I am now studying
whether this is doable in a reasonable amount of time.