Can now compute F(l_max) with lensing variance.
If you set 'compute_lensing_variance_lmax = yes' in the parameter
file, SONG will produce a 'fisher_lensvar_lmax.dat' file with the
Fisher matrix as a function of the maximum resolution of the considered
survey, including lensing variance. See 'fisher.h' for details on how this
is done, and why it is requires a separate option.