Introduced the concept of a data directory.
It is now possible to have separate directories for input
and output. The first is the 'data_directory' which is supposed
to contain the subfolders sources, transfers and bispectra. The
second is the 'run_directory' which will contain the parameter
files 'run_params.ini' and 'run_params.pre' plus the output
files (cl.dat, fisher.dat, etc.). By default the data directory
is taken to coincide with the run directory unless specified
differently in the parameter 'data_directory'.

The split between output run_dir and input data_dir is not strict as
the run_dir can also acts as an input (to read the parameter files when
a directory is given to SONG via the command line) and the data_dir
serves as output to create the sources, transfers and bispectra folders.

I know it is a mess. The whole I/O management needs a major overhaul.