First implementation of analytical ratios of 3j symbols.
Parity invariance requires even bispectra to vanish of odd values
of l1+l2+l3. This makes the interpolation of bispectra in l-space tricky.
For scalar bispectra (e.g. m=0 intrinsic bispectrum, primordial templates,
temperature CMB-lensing bispectrum), one can analytically factor out
the 3j symbol (l1,l2,l3)(0,0,0) to obtain a reduced bispectrum that is
defined also for odd values of l1+l2+l3.

For non-scalar bispectra (e.g. m!=0 intrinsic bispectrum, polarised
CMB-lensing), the extraction of (l1,l2,l3)(0,0,0) cannot be performed
analytically. In these cases, SONG would just adopt a grid
of even multipoles, so that l1+l2+l3 is always even and the 3j symbol
can be extracted numerically.

In this commit, instead, we use the analytical formulas from Schulten
& Gordon, 1961 (http://scitation.aip.org/content/aip/journal/jmp/16/10/10.1063/1.522426?ver=pdfcov)
to analytically extract the 3j-symbol (l1,l2,l3)(0,0,0).

So far we have included this feature as an option; to activate it, uncomment lines
2593-2603 in bispectra.c. It only works in a limited set of cases. In the future we will
implement it fully. This is important because it will allow us to compute
the intrinsic bispectrum for non-scalar configuration (m!=0) using an arbitrary l-grid.
That is, we will not have to choose an all-even grid of l, thus making interpolation of
the intrinsic bispectrum better behaved.