Started to add SONG support to CLASS submodule.
I have started to add SONG support to CLASS. CLASS is contained in SONG
as a submodule in the folder class.git, which points to the CLASS standalone
repository. The CLASS submodule is independent from SONG, but the reverse is
not true: SONG does track the commits in the CLASS submodule. If you commit
something on class.git, it will show up in song.git, but not viceversa.

As I update files in the class.git submodule, I will move the corresponding
files in song.git to a separate folder named `no_longer_needed`. Since the Makefile
won't look for source in that folder, SONG will automatically be compiled with
the corresponding (updated) files in the class.git subfolder. By the end of the process,
the song.git source folder will only contain SONG's modules (input2.c, perturbations2.c
etc.).

In this commit, I have added SONG support to the following files in class.git:
common.h, parser.c and parser.h. The corresponding files in SONG have been moved
to the folder `no_longer_needed`, which is ignored by the makefile.

Right now, SONG is broken, because all the other modules still rely on the
"old" common.h, parser.c and parser.h.