Streamlined computation of end of recombination in ppt2 module.
I have added the field index_tau_end_of_recombination in the perturbations2
structure, and the function perturb2_end_of_recombination() to compute it.

Nothing changes from the numerical point of view. The advantage is that
now index_tau_end_of_recombination is computed also when has_recombination_only
is false, in view of generalising the early/late ISW split.