Introduced effect switches for sw and isw in perturb2_sources().
Thanks to them and to the previous commit, it is now possible to
ask for only the early ISW effect at second order but still compute
the line of sight sources all the way to today. This is useful for
example when one wants to compute the effect of the quadratic potentials
(include_quad_metric_in_los_2nd_order=yes, which requires computation
all the way to today) but still needs to constrain the isw to early
times.