Implemented automatic r-sampling.
Now the user can choose the preferred method to sample the
integration variable r in the bispectrum integral, using the
variable bispectra_r_sampling:
 - If bispectra_r_sampling=custom, then the r will be sampled r_size
times between r_min and r_right.
 - If bispectra_r_sampling=centred, then the r will be sampled r_size
times between (tau0-tau_rec)-tau_rec*r_left and (tau0-tau_rec)+tau_rec*r_right.
This choice is justified by the fact that the transfer functions in the
bispectrum integral are peaked at recombination which, due to the
presence of the Bessel functions with argument k*r, makes the integrand
function peaked at r=tau0-tau.

The main advantage of this modification is that now the choice of r-sampling is
a purely numerical decision, independent of cosmology. That's why I have moved
the r-parameters from the bispectra.ini file to the precision files.