TCA1: first working version of quadrupoles.
I have coded in perturb2_save_early_transfers() the quadrupoles
in the tight coupling limit at first order in tau_c. In particular,
I have coded the photon quadrupole without needing to rely on any
other TCA quantity. This means that the TCA quadrupole is ready
to be implemented in SONG.