Implemented a variable starting time for differential system.
SONG now implements the same adaptive algorithm of CLASS to determine the start
time of the evolution the second-order differential system. The starting time
is determined using the same two criteria that used for the initial conditions,
namely:

  - The considered wavemode should still be in tight coupling regime.
  - The considered wavemode is still outside the horizon.

The strictness of the two criteria is controlled by the following parameters:

  - ppr2->start_small_k_at_tau_c_over_tau_h_song, which determines the start
    time using the ratio between the Compton interaction rate and the Hubble
    time.

  - ppr2->start_large_k_at_tau_h_over_tau_k_song, which determines the start
    time using the ratio between the Hubble time and the scale of the
    considered wavemode.

The smaller the parameters, the earlier the starting time.

I have also updated all the precision files to enable the option
for the variable starting time (tau_start_evolution=0) and that for
the tight coupling approximation (tight_coupling_approximation_song=1).