RSA: Conceptual commit.
I have understood a key point about the RSA approximation:

  If the radiation streaming approximation is turned on, we do not evolve
any equation for the massless hierarchies. The photon and neutrino
hierarchies will only include monopole and dipole, which will be inferred
with two constraint equation. Polarisation will be neglected altogether.

  This is the same approach as CLASS. But how does CLASS compute the effect of
reionisation on polarisation? Polarisation is sourced by the quadrupole, which
is neglected during RSA. Therefore, all modes that enter RSA before reionisation
(tau~4000), won't add up to polarisation. This behaviour is justified by the
fact that reionisation is affected by the modes that are the size of the horizon
at reionisation; the RSA however only affects subhorizon modes. Therefore, by
turning on the RSA we are not affecting reionisation at all, unless we make RSA
start earlier by reducing the radiation_streaming_trigger_tau_over_tau_k
parameter. After a quick test with CLASS v2.4.3, I find that the reionisation
contribution to C_l^EE is not affected as long as the parameter is larger than
~20.