RSA: added support for collisions during reionisation.
The RSA now approximately describes the perturbations even in the
presence of reionisation.

We found that an interesting result whereby the perturbed free-electron
fraction during and after reionisation is equal and opposite to the baryon
density contrast, thus cancelling the free-electron density contrast:
delta_e = delta_b + delta_xe ~ 0 for tau > tau_reio. If the computation
of delta_xe can be trusted, this means that at second-order and on
subhorizon scales (I haven't tested any other configuration yet) the
effect of reionisation is suppressed.

Note that we haven't tested yet whether the collisional RSA works
for modes where either k1 or k2 are superhorizon (basically squeezed
configurations).

Also:
* Included a quadratic correction to the RSA velocity of photons and neutrinos.
* Updated CLASS so that the first-order perturbations are written to file
correctly if SONG is called using the parameter k_output_values.

Now I have to:
* Check that the collisional RSA works also when either k1 or k2 are superhorizon.
* Check that the delta_e cancellation happens also when k is superhorizon.