Added sources time sampling in LOS integration grid.
If you set

  transfer2_tau_sampling = sources

in your parameter file, the line of sight integral will be solved
using an integration grid that matches the time sampling of the line
of sight sources (ppt2->tau_sampling). This option is faster because
there is no need for interpolation of the sources.

It seems that the results are the same using either

  transfer2_tau_sampling = sources

or

  transfer2_tau_sampling = smart
  tau_linstep_song = 0.6,

even though the second option puts many more grid points at late times.

I think the reason is that late-time effects such as ISW and reionisation
matter only for modes entering the horizon, which have a large wavelength
(k~2*pi/tau). For these modes, the time sampling of the soruce is dense enough,
because the oscillation frequency of the projection function J(k*(tau0-tau))
is proportional to k.