Implemented symmetric k sampling (by Christian Fidler)
It is a symmetric sampling in (k1,k2,k3) by means of a transformation whereby
the three wavemodes end up sharing the same grid. This sampling allows to describe
the physical Fourier space as a cube despite the triangular condition.

The symmetric sampling is crucial to compute the power spectrum P(k) of second-order
perturbations, as it allows to sample the edges of the triangular condition with a
logarithmic power.

SONG's default sampling, the 'smart' sampling, will always fail to describe P(k) at
some small enough value of k. In some cases (eg. the magnetic field at early times), the
failure happens quicker because of a specific structure of the kernel, but also for
the simple power spectrum today you need the symmetric sampling at small k. Note that the
symmetric sampling can be considered the Fourier space version of the (l1,l2,l3)
sampling adopted in Ferguson et al 2009 (http://arxiv.org/abs/0812.3413) for the cubic
interpolation of the bispectrum.