Working version of 2nd-order P(k) computation (Christian Fidler)
If you run

./song ini/matter_pks.ini pre/matter_pks.pre

SONG will compute the P(k) for the magnetic fields and for the CDM
density at second order. The result is not accurate yet because I
have not implemented the symmetric sampling in spectra2.c, but it
should be not too far from the true P(k).

Note that in spectra2_integrate_fourier() I have implemented two
alternative implementation of step size for the P(k) integration,
which give quite different results at high k.